+++ wtt/src/commands.rs
@@ -1,10 +1,12 @@
 mod add;
+mod gc;
 mod list;
 mod remove;
 mod setup;
 mod teardown;
 
 pub use add::Add;
+pub use gc::Gc;
 pub use list::List;
 pub use remove::Remove;
 pub use setup::Setup;